用户，商品，管理员接口初步调试完成 //接口已通过-OK就代表接口没问题

+++ Mall-ShopX/Mall-pojo/src/main/java/com/itcast/admin/entity/DityUrl.java
@@ -16,10 +16,10 @@ import lombok.NoArgsConstructor;
 public class DityUrl {
 
     @ApiModelProperty("主键值")
-    private Long id;
+    private int id;
 
     @ApiModelProperty("关联商品id")
-    private Long dityId;
+    private int dityId;
 
     @ApiModelProperty("图片路径")
     private String avatar;
